# public/password-reset-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Password Reset API",
    "description": "Complete Password Reset API Documentation with Email Link and OTP Methods",
    "version": "1.0.0",
    "contact": {
      "name": "Medical Courier Support",
      "email": "support@medicalcourier.com"
    }
  },
  "servers": [
    {
      "url": "http://172.16.123.158:8000",
      "description": "API Server"
    }
  ],
  "tags": [
    {
      "name": "Password Reset",
      "description": "Password reset endpoints with email link and OTP methods for both web and mobile applications"
    }
  ],
  "paths": {
    "/api/v1/forgot-password": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Send password reset link via email",
        "description": "Sends a secure password reset link to the user's email address. The link expires in 60 minutes. Designed for web applications.",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "description": "User email address",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "driver@test.com",
                    "description": "Registered email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password reset link sent to your email"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "If your email exists in our system, you will receive a password reset link shortly."
                        },
                        "dev_token": {
                          "type": "string",
                          "nullable": true,
                          "example": "a1b2c3d4e5f6...",
                          "description": "Reset token (only in development environment)"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/password/send-otp": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Send OTP code for password reset (Mobile)",
        "description": "Sends a 6-digit OTP code to the user's email for password reset. Designed for mobile apps. OTP expires in 15 minutes.",
        "operationId": "sendOtp",
        "requestBody": {
          "required": true,
          "description": "User email address",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "driver@test.com",
                    "description": "Registered email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OTP sent to your email"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "If your email exists in our system, you will receive an OTP code shortly."
                        },
                        "expires_in_minutes": {
                          "type": "integer",
                          "example": 15
                        },
                        "dev_otp": {
                          "type": "string",
                          "nullable": true,
                          "example": "123456",
                          "description": "OTP code (only in development environment)"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/password/verify-otp": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Verify OTP code (Optional Step)",
        "description": "Verifies the OTP code before password reset. This is an optional step - you can skip directly to reset-with-otp endpoint.",
        "operationId": "verifyOtp",
        "requestBody": {
          "required": true,
          "description": "Email and OTP code",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "driver@test.com"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456",
                    "description": "6-digit OTP code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "OTP verified successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "verified": {
                          "type": "boolean",
                          "example": true
                        },
                        "message": {
                          "type": "string",
                          "example": "OTP verified successfully. You can now reset your password."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reset-password": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Reset password using token from email link",
        "description": "Resets user password using the token received in the email reset link. Token expires in 60 minutes.",
        "operationId": "resetPassword",
        "requestBody": {
          "required": true,
          "description": "Password reset data with token",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "token", "password", "password_confirmation"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "driver@test.com"
                  },
                  "token": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6...",
                    "description": "Reset token from email link"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassword123",
                    "description": "New password (min 8 characters)"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassword123",
                    "description": "Confirm new password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password reset successfully. You can now login with your new password."
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/password/reset-with-otp": {
      "post": {
        "tags": ["Password Reset"],
        "summary": "Reset password using OTP code (Mobile)",
        "description": "Resets user password using the 6-digit OTP code received via email. Designed for mobile apps. OTP expires in 15 minutes.",
        "operationId": "resetPasswordWithOtp",
        "requestBody": {
          "required": true,
          "description": "Password reset data with OTP",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "otp", "password", "password_confirmation"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "driver@test.com"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456",
                    "description": "6-digit OTP code from email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassword123",
                    "description": "New password (min 8 characters)"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassword123",
                    "description": "Confirm new password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password reset successfully. You can now login with your new password."
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation error"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": ["The email field is required."],
              "password": ["The password must be at least 8 characters."]
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid or expired OTP. Please request a new one."
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  }
}